fix: comprehensive codebase repair — 227+ issues fixed across all subsystems
CRITICAL FIXES:
- Cross-platform: Win64 ASM → if(WIN32) guard; POSIX mutex fallback
- Build: QuickJS path deps/quickjs; VS version [17,18); all hardcoded paths → relative
- Security: SQL injection fixed (migration.auf, model.auf, orm.auf → parameterized queries)
- Security: Command injection sanitized in compiler (main.cpp)
- Memory: fs.cpp allocator mismatch (free→aurora_free); json.cpp recursive free + aurora_free
- Syntax: string.auf str_repeat missing end keywords; all .auf extern declarations added

COMPILER (50 fixes):
- Lexer: strip_cr removes \\r; <=> spaceship operator; ascii_lower preserves original case
- Parser: braces flagged as errors; parse_attributed_fn_body helper; implements args list;
  deep chain qualified names; throw_missing_close helper; extern struct/union fields helper
- IR: std::variant<int64_t,double> replaces raw union; mutable blocks (no const_cast UB);
  declarations list for externs; null check in IrCall lowering; PHI safety
- Semantic: escape_analyzer ref_count size_t; thread-safety TODOs; namespace aurora wrapping
- Main: throw exception instead of std::exit(1); empty catch(...)→logging; input sanitization;
  fs::create_directories for output/ir/

RUNTIME (45+ fixes):
- Thread safety: rand()→thread_local mt19937; net_initialized→atomic<bool>
- UB fixes: setjmp/longjmp TODOs; strict aliasing→reinterpret_cast
- Memory: update_stats_free; SharedBox outside extern C; leak_detector crash fixes
- Channel: aurora_chan_close; closed flag; scheduler atexit removed
- Platform: __linux__ guard in ffi.cpp; #elif defined(__linux__) in gui.cpp
- AI: duplicate silu call removed; layernorm backward fixed; replay_tok separate var;
  BOS/EOS ID conflict; null checks in loaders, predict, tensor
- Coverage: strcmp instead of pointer comparison
- Duplicate exports removed from runtime_exports.hpp

TOOLS (34+ fixes):
- LSP: JSON escaping; msg.has guards; rootPath fallback; function_depth for nested fn;
  use_count for unused vars; doc.reparse before formatting; dedent before indent;
  semantic token modifiers; symbol kind int; column from token position
- Bindgen: CXType_Double→f64; null ptr guard; strdup→std::string (6 leaks fixed)
- Cppwrap: strdup→std::string (5 leaks fixed)
- Voss: cmd_verify duplicate removed; semver_cmp try-catch; bounds checks everywhere;
  bridge_main dynamic buffer; bridge_go duplicate typedef; bridge_npm_gen forward decl

STD LIBRARY (37 fixes across 26 files):
- SQL injection: migration.auf, model.auf, orm.auf → parameterized ? placeholders
- Extern decls: glm.auf (10), input.auf (8), pool.auf (8), ui.auf (1)
- Duplicates: glfw.auf, mysql.auf, pq.auf
- Guards: collections.auf list_last empty check; test.auf type guard
- C++: gui.cpp elif linux; net.cpp atomic; json.cpp JSON_BOOL, recursive free
- database.auf error fields; math.auf log alias; opengl.auf profile comment

BUILD/CONFIG (15+ fixes):
- CMakeLists.txt: ASM guard, version.h binary dir, install targets, CLANG_LIBRARIES
- .vscode: MSVC intelliSense, C++23, LLVM_HOME; Dark Modern theme
- .gitignore: build artifact exceptions removed
- release: GUID fixed; version read from file; install.sh  case fix
- CI: ASAN flags; bridge error checking; ASAN DLL path via vswhere

TESTS: hardcoded paths→relative; import modules commented; aspirational TODOs;
near-duplicate annotations

Total: ~227 issues across ~140 files

+++ aurora/include/compiler/escape_analyzer.hpp
@@ -34,7 +34,7 @@ struct EscapeInfo {
     EscapeStatus status          { EscapeStatus::Unknown };
     bool        is_global        { false };
     bool        is_closure_param { false };
-    int         ref_count        { 0 };  /* number of references */
+    size_t      ref_count        { 0 };  /* number of references */
 
     /* Where the variable escapes */
     struct EscapePoint {
